+++ test/unit/lib/models/Log.mocks.js
@@ -11,7 +11,13 @@ const Mock = function () {
   }
 
   const LogStub = sandbox.stub().returns(logStub)
-  LogStub.find = sandbox.stub().usingPromise().resolves()
+  const findSort = sandbox.stub().usingPromise().resolves()
+  LogStub.findSort = findSort
+  LogStub.find = sandbox.stub().usingPromise().returns({
+    sort: findSort
+  })
+  LogStub.countDocuments = sandbox.stub().usingPromise().resolves()
+  LogStub.estimatedDocumentCount = sandbox.stub().returns(4)
 
   const stubs = {
     log: logStub,